move RebalancePlanGenerator to generator package
